findbugs: fix problems
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@51002 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ web/src/main/java/org/geogebra/web/web/gui/view/algebra/LatexTreeItem.java
@@ -382,7 +382,7 @@ public class LatexTreeItem extends RadioTreeItem {
 
 
 	@Override
-	public void setPixelRatio(float pixelRatio) {
+	public void setPixelRatio(double pixelRatio) {
 		if (mf != null) {
 			mf.setPixelRatio(pixelRatio);
 			mf.repaint();
